signed changes?

+++ yolov7/deploy/triton-inference-server/processing.py
@@ -28,7 +28,7 @@ def preprocess(img, input_shape, letter_box=True):
 def postprocess(num_dets, det_boxes, det_scores, det_classes, img_w, img_h, input_shape, letter_box=True):
     boxes = det_boxes[0, :num_dets[0][0]] / np.array([input_shape[0], input_shape[1], input_shape[0], input_shape[1]], dtype=np.float32)
     scores = det_scores[0, :num_dets[0][0]]
-    classes = det_classes[0, :num_dets[0][0]].astype(np.int)
+    classes = det_classes[0, :num_dets[0][0]].astype(np.int32)
 
     old_h, old_w = img_h, img_w
     offset_h, offset_w = 0, 0
@@ -43,7 +43,7 @@ def postprocess(num_dets, det_boxes, det_scores, det_classes, img_w, img_h, inpu
     boxes = boxes * np.array([old_w, old_h, old_w, old_h], dtype=np.float32)
     if letter_box:
         boxes -= np.array([offset_w, offset_h, offset_w, offset_h], dtype=np.float32)
-    boxes = boxes.astype(np.int)
+    boxes = boxes.astype(np.int32)
 
     detected_objects = []
     for box, score, label in zip(boxes, scores, classes):
